pyproject.toml
[project]
name = "job-application-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gspread>=6.2.1",
    "ipykernel>=7.1.0",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
